Web Application to Command Line Application. Additional criteria to fetch pending transactions. New service method to handle receipt with single realisation into bank.

+++ src/main/java/org/sakuram/persmony/repository/DomainValueRepository.java
@@ -3,6 +3,6 @@ package org.sakuram.persmony.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.sakuram.persmony.bean.DomainValue;
 
-public interface DomainValueRepository extends JpaRepository<DomainValue, Integer> {
+public interface DomainValueRepository extends JpaRepository<DomainValue, Long> {
 	
 }
