+++ src/main/java/org/sakuram/persmony/repository/InvestmentTransactionRepository.java
@@ -3,6 +3,6 @@ package org.sakuram.persmony.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.sakuram.persmony.bean.InvestmentTransaction;
 
-public interface InvestmentTransactionRepository extends JpaRepository<InvestmentTransaction, Integer> {
+public interface InvestmentTransactionRepository extends JpaRepository<InvestmentTransaction, Long> {
 	
 }
