+++ src/main/java/org/sakuram/persmony/repository/RealisationRepository.java
@@ -3,6 +3,6 @@ package org.sakuram.persmony.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.sakuram.persmony.bean.Realisation;
 
-public interface RealisationRepository extends JpaRepository<Realisation, Integer> {
+public interface RealisationRepository extends JpaRepository<Realisation, Long> {
 	
 }
