+++ src/main/java/org/sakuram/persmony/repository/SavingsAccountTransactionRepository.java
@@ -3,6 +3,6 @@ package org.sakuram.persmony.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.sakuram.persmony.bean.SavingsAccountTransaction;
 
-public interface SavingsAccountTransactionRepository extends JpaRepository<SavingsAccountTransaction, Integer> {
+public interface SavingsAccountTransactionRepository extends JpaRepository<SavingsAccountTransaction, Long> {
 	
 }
